requirements.txt
# Core requirements
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Jupyter notebooks
jupyter>=1.0.0
notebook>=6.4.0

# YARA integration
yara-python>=4.2.0

# Model persistence
joblib>=1.1.0

# Utility
tqdm>=4.62.0

# Testing
pytest>=6.2.5